refactoring grid anatomy
ok, we are going through a major structure change for grid. I want to limit the role=grid to only cover grid header, grid body and grid footer, focus trapping will only work inside that role=grid component. Maybe we can name it GridList.

structure:
GridList
  GridHeader
  GridBody
    GridRow
  GridFooter

the Caption and Title will be adjacent to that GridList and inside a GridListContainer

GridListContainer
  GridTItle
  GridCaption
  GridList

GridListContainer main job is to distribute the grid id automatically between title, caption and the role=grid component itself and it's optional for the opration of a grid list. Without container, the gridlist can accept an id manually.

However, when a GridListContainer is present, the gridlist state (selection, focused, etc.) will contextual to the entire GridListContainer. Otherwise, it's only stored within the GridList component.

GridListContainer is expected to contain only one GridList, if there are more than one, they will share the same state and technically a mirrored version of each other.

Signed-off-by: Tung Vu <me@tungv.com>